# src/Web/ClientApp/flutter/openapi3_v1.yaml
...
openapi: 3.0.0
info:
  title: Mac.CarHub API
  version: 1.0.0
paths:
...
  /api/Authentication/reset-password:
    post:
      tags:
        - Authentication
      operationId: ResetPassword
      parameters:
        - name: Accept-Language
          in: header
          description: Language preference for the response.
          schema:
            type: string
            items:
              type: string
            enum:
              - ar-EG
              - en-US
      requestBody:
        x-name: resetRequest
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
        required: true
        x-position: 1
      responses:
        '200':
          description: ''
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpValidationProblemDetails'
...
components:
  schemas:
    HttpValidationProblemDetails:
      allOf:
        - $ref: '#/components/schemas/ProblemDetails'
        - type: object
          additionalProperties:
            nullable: true
          properties:
            errors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
    ProblemDetails:
      type: object
      additionalProperties:
        nullable: true
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
...
    ResetPasswordRequest:
      type: object
      additionalProperties: false
      properties:
        email:
          type: string
        resetCode:
          type: string
        newPassword:
          type: string
...
  securitySchemes:
    XApiKey:
      type: apiKey
      description: 'Type into the text-box: X-Api-Key {your API key}.'
      name: X-Api-Key
      in: header
    JWT:
      type: apiKey
      description: 'Type into the text-box: Bearer {your JWT token}.'
      name: Authorization
      in: header
security:
  - XApiKey: []
  - JWT: []